tweak(server/controllers): rename user.controller to follow camelCase as other file names
